Split duplicative code to external function
